Major steps forward towards high level API for hybridized FE solvers
low-level HDG tests are now broken. The new overloads of
IntegrationMap() lead to the application of DensifyInnerMostBlock() map
to a block array which does not have the required blocked information.
I am positive that this issue will be by-passed with the high level
api as we will apply DensifyInnerMostBlock() to the sum of several
terms