Solved the following:
Some of the terms of the bilinear form of the LDG-H method, for some particular orderings of the operations in the integrand, lead to situations in which one has to perform an operation among a CellField defined on the bulk and a CellField defined on the skeleton (or vice-versa). This requires a change_domain transformation which is not available yet (pending). Thus, at present, we have to force a particular order of evaluation in the bilinear form terms in order to bypass the lack of this piece of sw.