Completed implementation of stress_assisted_diffusion_hencky. In particular, I have considered two implementation approaches:   (1) Solve the Jacobian hybridizable system directly       (via standard Gridap FEOperator).   (2) Hybridize Jacobian system (via a newly developed type       HybridFEOperator)
(2) is not robust (crashes), and has to be debugged.
This is the way to go with efficiency in mind.

(1) does not crash, runs bottom-top. However, it is not even able
to solve a linear problem with manufactured solution in the FE space.
The Newton method converges in one iteration, but not to the correct
solution. We need to identify potential causes of error in
the formulation itself and/or the code.