Started to play with 2-level V MG cycle for HDG + linear elasticity + strong symmetric stresses
I need to perform more systematic numerical experiments, but there
seems so far to be a very mild degradation in # iterations with h (with Darcy
it reached asymptotic regime)

When deactivating the projection Pm, there is still mild degradation
in # iterations, but seems to be milder than with Pm.

Finally, I am trying to precondition the operator corresponding to k=0.
We know that with k=0 the formulation does not convergence to the true
solution. Thus, not sure if this result is useful for preconditioning.